+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/communication/chat/ChatThreadClient.java
@@ -22,8 +22,6 @@ import com.azure.core.http.rest.Response;
 import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
 
-import java.util.Collections;
-
 /**
  * Sync Client that supports chat thread operations.
  */