+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/communication/chat/models/InvalidParticipantException.java
@@ -5,15 +5,28 @@ package com.azure.communication.chat.models;
 
 import com.azure.core.exception.AzureException;
 
-public class InvalidParticipantException extends AzureException {
+/**
+ * Exception when a participant cannot be added to a chat thread
+ */
+public final class InvalidParticipantException extends AzureException {
 
-    private ChatError chatError;
+    private transient ChatError chatError;
 
+    /**
+     * Constructs a new InvalidParticipantException
+     *
+     * @param chatError the ChatError underlying this exception
+     */
     public InvalidParticipantException(ChatError chatError) {
         super(chatError.getMessage());
         this.chatError = chatError;
     }
 
+    /**
+     * Gets the underlying ChatError returned from the server
+     *
+     * @return the CHatError
+     */
     public ChatError getChatError() {
         return chatError;
     }