+++ sdk/communication/azure-communication-chat/CHANGELOG.md
@@ -13,7 +13,7 @@
 - Removed `AddChatParticipantsOptions`
 - Changed `addParticipants` to take `Iterable<ChatParticipant>` instead of `AddChatParticipantsOptions`
 - Added `context` parameter to the max overloads of `listParticipants`, `listReadReceipts`
-
+- `CreateChatThreadOptions` constructor now requires `topic`
 
 ## 1.0.0-beta.6 (2021-03-09)
 Updated `azure-communication-chat` version